[LV] Move isCandidateForEpilogueVectorization to LoopVectorizationPlanner.cpp (NFC).
isCandidateForEpilogueVectorization (together with its file-static
helper hasUnsupportedHeaderPhiRecipe) does not depend on anything in
LoopVectorize.cpp, move it to LoopVectorizationPlanner.cpp.
